chat broke down to composables & components
