+++ src/components/ChatInput.vue
@@ -41,5 +41,51 @@ defineExpose({ messageInput })
 </script>
 
 <style scoped>
-/* Copy relevant styles from original CSS */
+textarea {
+  flex: 1;
+  border: none;
+  resize: none;
+  padding: 5px;
+  font-family: inherit;
+  font-size: inherit;
+  outline: none;
+}
+
+.input-area {
+  border-top: 1px solid #e5e5e5;
+  padding: 20px;
+  max-width: 800px;
+  margin: 0 auto;
+  width: 100%;
+}
+
+.input-container {
+  display: flex;
+  gap: 10px;
+  background-color: #ffffff;
+  border: 1px solid #e5e5e5;
+  border-radius: 10px;
+  padding: 10px;
+}
+
+.send-btn {
+  background-color: #b1a184;
+  color: white;
+  border: none;
+  border-radius: 5px;
+  padding: 5px 15px;
+  cursor: pointer;
+}
+
+.send-btn:disabled {
+  background-color: #eae3d6;
+  cursor: not-allowed;
+}
+
+.disclaimer {
+  text-align: center;
+  color: #666;
+  font-size: 0.8em;
+  margin-top: 10px;
+}
 </style>
